fix metadata covers
